// src/bindings/Schema.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Routine } from "./Routine";
import type { Sequence } from "./Sequence";
import type { Table } from "./Table";
import type { UserType } from "./UserType";

export type Schema = { name: string, tables: Array<Table>, routines: Array<Routine>, 
/**
 * PostgreSQL's; BigQuery has none.
 */
sequences: Array<Sequence>, 
/**
 * The types made with `CREATE TYPE` or `CREATE DOMAIN`, PostgreSQL's.
 */
types: Array<UserType>, };